+++ modules/auth/serivices.js
@@ -1,3 +1,12 @@
+const UserService = require('../user/serivices');
+
+const userService = new UserService();
+
 class AuthService {
-    public 
+    async login(email, password) {
+        const validateUser = await userService.login(email, password);
+        return validateUser;
+    }
 }
+
+module.exports = AuthService;