+++ modules/user/controller.js
@@ -1,10 +1,12 @@
 const UserService = require('./serivices');
+const ProfileService = require('../profile/serivices');
 const response = require('../../utils/responses');
 const { v4: uuidv4 } = require('uuid');
 const { hashPassword } = require('../../utils/password.helper');
 const { getCurrentDateUTC } = require('../../utils/date.helper');
 
 const service = new UserService();
+const profileService = new ProfileService();
 
 class UserController {
     constructor () {};
@@ -19,7 +21,8 @@ class UserController {
                 password: await hashPassword(password),
                 createdAt: getCurrentDateUTC()
             }
-            await service.createUser(user);
+            const newUser = await service.createUser(user);
+            await profileService.createProfile(newUser.dataValues.id)
             response.success(req, res, [], 201);
         }catch(error){
             response.error(req, res, error.message, 500);